+++ docs/src/examples/components/Button/Types/ButtonExampleEmphasis.shorthand.tsx
@@ -4,8 +4,7 @@ import { Button } from '@stardust-ui/react'
 const ButtonExampleEmphasis = () => (
   <div>
     <Button type="primary" content="Primary" />
-    <Button type="secondary" content="Secondary" />
-    <Button type="secondaryTinted" content="Secondary Tinted" />
+    <Button content="Normal/Default" />
   </div>
 )