+++ docs/src/examples/components/Button/Types/ButtonExampleEmphasis.tsx
@@ -4,8 +4,7 @@ import { Button } from '@stardust-ui/react'
 const ButtonExampleEmphasis = () => (
   <div>
     <Button type="primary">Primary</Button>
-    <Button type="secondary">Secondary</Button>
-    <Button type="secondaryTinted">Secondary Tinted</Button>
+    <Button>Normal/Default</Button>
   </div>
 )