+++ src/themes/teams-dark/components/Button/buttonVariables.ts
@@ -2,6 +2,9 @@ export interface IButtonVariables {
   [key: string]: string | number
 
   color: string
+  colorActive: string
+  colorHover: string
+  colorFocus: string
   colorDisabled: string
   backgroundColor: string
   backgroundColorActive: string
@@ -26,33 +29,19 @@ export interface IButtonVariables {
   primaryBorderColorFocus: string
   primaryBorderColorFocusIndicator: string
 
-  secondaryTintedColor: string
-  secondaryTintedColorActive: string
-  secondaryTintedColorHover: string
-  secondaryTintedColorFocus: string
-  secondaryTintedBackgroundColor: string
-  secondaryTintedBackgroundColorActive: string
-  secondaryTintedBackgroundColorHover: string
-  secondaryTintedBackgroundColorFocus: string
-  secondaryTintedBorderColor: string
-  secondaryTintedBorderColorActive: string
-  secondaryTintedBorderColorHover: string
-  secondaryTintedBorderColorFocus: string
-  secondaryTintedBorderColorFocusIndicator: string
-
   primaryCircularBorderColorFocusIndicator: string
 
-  secondaryCircularColor: string
-  secondaryCircularColorActive: string
-  secondaryCircularBackgroundColor: string
-  secondaryCircularBackgroundColorActive: string
-  secondaryCircularBackgroundColorHover: string
-  secondaryCircularBackgroundColorFocus: string
-  secondaryCircularBorderColor: string
-  secondaryCircularBorderColorActive: string
-  secondaryCircularBorderColorHover: string
-  secondaryCircularBorderColorFocus: string
-  secondaryCircularBorderColorFocusIndicator: string
+  circularColor: string
+  circularColorActive: string
+  circularBackgroundColor: string
+  circularBackgroundColorActive: string
+  circularBackgroundColorHover: string
+  circularBackgroundColorFocus: string
+  circularBorderColor: string
+  circularBorderColorActive: string
+  circularBorderColorHover: string
+  circularBorderColorFocus: string
+  circularBorderColorFocusIndicator: string
 
   textColorHover: string
   textPrimaryColor: string
@@ -66,6 +55,9 @@ export interface IButtonVariables {
 export default (siteVars: any): IButtonVariables => {
   return {
     color: siteVars.white,
+    colorActive: siteVars.white,
+    colorHover: siteVars.white,
+    colorFocus: siteVars.white,
     colorDisabled: siteVars.gray06,
     backgroundColor: 'transparent',
     backgroundColorActive: siteVars.gray08,
@@ -90,33 +82,19 @@ export default (siteVars: any): IButtonVariables => {
     primaryBorderColorFocus: siteVars.black,
     primaryBorderColorFocusIndicator: siteVars.white,
 
-    secondaryTintedColor: siteVars.brand06,
-    secondaryTintedColorActive: siteVars.brand04,
-    secondaryTintedColorHover: siteVars.brand04,
-    secondaryTintedColorFocus: siteVars.brand02,
-    secondaryTintedBackgroundColor: siteVars.black,
-    secondaryTintedBackgroundColorActive: siteVars.brand14,
-    secondaryTintedBackgroundColorHover: siteVars.brand16,
-    secondaryTintedBackgroundColorFocus: siteVars.brand14,
-    secondaryTintedBorderColor: siteVars.brand14,
-    secondaryTintedBorderColorActive: siteVars.brand12,
-    secondaryTintedBorderColorHover: siteVars.brand12,
-    secondaryTintedBorderColorFocus: siteVars.black,
-    secondaryTintedBorderColorFocusIndicator: siteVars.brand02,
-
     primaryCircularBorderColorFocusIndicator: siteVars.white,
 
-    secondaryCircularColor: siteVars.gray02,
-    secondaryCircularColorActive: siteVars.black,
-    secondaryCircularBackgroundColor: 'transparent',
-    secondaryCircularBackgroundColorActive: siteVars.gray02,
-    secondaryCircularBackgroundColorHover: siteVars.gray03,
-    secondaryCircularBackgroundColorFocus: siteVars.gray02,
-    secondaryCircularBorderColor: siteVars.gray02,
-    secondaryCircularBorderColorActive: 'transparent',
-    secondaryCircularBorderColorHover: 'transparent',
-    secondaryCircularBorderColorFocus: 'transparent',
-    secondaryCircularBorderColorFocusIndicator: siteVars.black,
+    circularColor: siteVars.gray02,
+    circularColorActive: siteVars.black,
+    circularBackgroundColor: 'transparent',
+    circularBackgroundColorActive: siteVars.gray02,
+    circularBackgroundColorHover: siteVars.gray03,
+    circularBackgroundColorFocus: siteVars.gray02,
+    circularBorderColor: siteVars.gray02,
+    circularBorderColorActive: 'transparent',
+    circularBorderColorHover: 'transparent',
+    circularBorderColorFocus: 'transparent',
+    circularBorderColorFocusIndicator: siteVars.black,
 
     textColorHover: siteVars.brand04,
     textPrimaryColor: siteVars.brand,